% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-model.R
\docType{class}
\name{cache_model_class-class}
\alias{cache_model_class-class}
\title{In-Memory backend that simply stores everything in memory and matrix form}
\description{
In-Memory backend that simply stores everything in memory and matrix form
}
\keyword{internal}
