% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-model.R
\docType{class}
\name{cache_model_class-class}
\alias{cache_model_class-class}
\alias{moi_supports,cache_model_class,MOI_variable_index_attribute-method}
\alias{moi_get,cache_model_class,MOI_variable_index_attribute,character-method}
\alias{moi_supports,cache_model_class,MOI_variable_name_attribute-method}
\alias{moi_set,cache_model_class,MOI_variable_name_attribute,MOI_variable_index,character-method}
\alias{moi_set,cache_model_class,MOI_objective_function,MOI_scalar_affine_function,missing-method}
\alias{moi_set,cache_model_class,MOI_objective_sense,numeric,missing-method}
\alias{moi_add_variable,cache_model_class-method}
\alias{moi_add_constraint,cache_model_class,MOI_abstract_function,MOI_abstract_set-method}
\alias{moi_add_constraint,cache_model_class,MOI_abstract_index,MOI_abstract_set-method}
\alias{moi_add_constrained_variable,cache_model_class,MOI_abstract_set-method}
\title{In-Memory backend that simply stores everything in-memory}
\usage{
\S4method{moi_supports}{cache_model_class,MOI_variable_index_attribute}(model, type)

\S4method{moi_get}{cache_model_class,MOI_variable_index_attribute,character}(model, type, index)

\S4method{moi_supports}{cache_model_class,MOI_variable_name_attribute}(model, type)

\S4method{moi_set}{cache_model_class,MOI_variable_name_attribute,MOI_variable_index,character}(model, type, index, value)

\S4method{moi_set}{cache_model_class,MOI_objective_function,MOI_scalar_affine_function,missing}(model, type, index, value)

\S4method{moi_set}{cache_model_class,MOI_objective_sense,numeric,missing}(model, type, index, value)

\S4method{moi_add_variable}{cache_model_class}(model)

\S4method{moi_add_constraint}{cache_model_class,MOI_abstract_function,MOI_abstract_set}(model, func, set)

\S4method{moi_add_constraint}{cache_model_class,MOI_abstract_index,MOI_abstract_set}(model, func, set)

\S4method{moi_add_constrained_variable}{cache_model_class,MOI_abstract_set}(model, set)
}
\description{
In-Memory backend that simply stores everything in-memory

Add a constraint variable to a cached model
}
\keyword{internal}
