% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{canonicalize}
\alias{canonicalize}
\alias{canonicalize,MOI_scalar_affine_function-method}
\alias{canonicalize,MOI_scalar_quadratic_function-method}
\title{Bring object to a canonical form}
\usage{
canonicalize(x)

\S4method{canonicalize}{MOI_scalar_affine_function}(x)

\S4method{canonicalize}{MOI_scalar_quadratic_function}(x)
}
\arguments{
\item{x}{an object}
}
\description{
Bring object to a canonical form
}
