% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-attributes.R
\docType{data}
\name{FEASIBILITY_SENSE}
\alias{FEASIBILITY_SENSE}
\title{Optimization feasibility sense attribute}
\format{
An object of class \code{MOI_optimization_sense_feasibility} of length 1.
}
\usage{
FEASIBILITY_SENSE
}
\description{
Optimization feasibility sense attribute
}
\keyword{datasets}
