% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{interval_set}
\alias{interval_set}
\title{An instance of an interval set}
\usage{
interval_set(lower, upper)
}
\arguments{
\item{lower}{a length 1 numeric vector}

\item{upper}{a length 1 numeric vector}
}
\description{
An instance of an interval set
}
