% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{less_than_set}
\alias{less_than_set}
\title{An instance of a less than or equal set}
\usage{
less_than_set(upper)
}
\arguments{
\item{upper}{a length 1 numeric vector}
}
\description{
An instance of a less than or equal set
}
