% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-attributes.R
\docType{data}
\name{MAX_SENSE}
\alias{MAX_SENSE}
\title{Optimization max attribute}
\format{
An object of class \code{MOI_optimization_sense_max} of length 1.
}
\usage{
MAX_SENSE
}
\description{
Optimization max attribute
}
\keyword{datasets}
