% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-attributes.R
\docType{class}
\name{MOI_abstract_constraint_attribute-class}
\alias{MOI_abstract_constraint_attribute-class}
\title{An abstract constraint attribute}
\description{
An abstract constraint attribute
}
