% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indextypes.R
\docType{class}
\name{MOI_abstract_index-class}
\alias{MOI_abstract_index-class}
\title{An abstract index}
\description{
An abstract index
}
\section{Slots}{

\describe{
\item{\code{value}}{a length one integer like numeric}
}}

