% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_add_constrained_variable,MOI_abstract_model,MOI_abstract_set-method}
\alias{moi_add_constrained_variable,MOI_abstract_model,MOI_abstract_set-method}
\title{Add a constraint variable}
\usage{
\S4method{moi_add_constrained_variable}{MOI_abstract_model,MOI_abstract_set}(model, set)
}
\arguments{
\item{model}{a model}

\item{set}{a set}
}
\description{
Add a constraint variable
}
