% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_add_constrained_variable}
\alias{moi_add_constrained_variable}
\title{Add a constrained variable}
\usage{
moi_add_constrained_variable(model, set)
}
\arguments{
\item{model}{a MOI model}

\item{set}{the constraint set of the variable}
}
\description{
Add a constrained variable
}
