% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_add_constraint}
\alias{moi_add_constraint}
\title{Add a constraint to a model}
\usage{
moi_add_constraint(model, func, set)
}
\arguments{
\item{model}{the model}

\item{func}{a function in the MOI sense (e.g. a MOI_affine_term)}

\item{set}{the constraint set}
}
\description{
Add a constraint to a model
}
