% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_add_variables,MOI_abstract_model,numeric-method}
\alias{moi_add_variables,MOI_abstract_model,numeric-method}
\title{Batch add a number of variables to a model}
\usage{
\S4method{moi_add_variables}{MOI_abstract_model,numeric}(model, n_vars)
}
\arguments{
\item{model}{the model}

\item{n_vars}{the number of variables}
}
\description{
Batch add a number of variables to a model
}
