% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_add_variables}
\alias{moi_add_variables}
\title{Add variables}
\usage{
moi_add_variables(model, n_vars)
}
\arguments{
\item{model}{a MOI model}

\item{n_vars}{a integer like numeric >= 1}
}
\description{
Add variables
}
