% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-attributes.R
\docType{class}
\name{MOI_constraint_dual-class}
\alias{MOI_constraint_dual-class}
\title{A dual of a constraint}
\description{
A dual of a constraint
}
\section{Slots}{

\describe{
\item{\code{result_idx}}{a length 1 numeric}
}}

