% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-attributes.R
\name{moi_constraint_dual}
\alias{moi_constraint_dual}
\title{A new constraint dual}
\usage{
moi_constraint_dual(result_idx = 1)
}
\arguments{
\item{result_idx}{a scalar index >= 1}
}
\description{
A new constraint dual
}
