% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indextypes.R
\docType{class}
\name{MOI_constraint_index-class}
\alias{MOI_constraint_index-class}
\title{A constraint index}
\description{
A constraint index
}
\section{Slots}{

\describe{
\item{\code{value}}{a length one integer like numeric}
}}

