% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{class}
\name{MOI_dual_status-class}
\alias{MOI_dual_status-class}
\title{The dual status attribute of an optimizer}
\description{
The dual status attribute of an optimizer
}
\section{Slots}{

\describe{
\item{\code{result_idx}}{the length 1 index}
}}

