% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\name{moi_dual_status}
\alias{moi_dual_status}
\title{The dual status attribute}
\usage{
moi_dual_status(result_idx = 1)
}
\arguments{
\item{result_idx}{a length 1 scalar >= 1}
}
\description{
The dual status attribute
}
