% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\docType{class}
\name{MOI_equal_to_set-class}
\alias{MOI_equal_to_set-class}
\title{An equal-to set}
\description{
An equal-to set
}
\section{Slots}{

\describe{
\item{\code{value}}{A length-one numeric vector}
}}

