% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{moi_equal_to_set}
\alias{moi_equal_to_set}
\title{An instance of an equal-to set}
\usage{
moi_equal_to_set(value)
}
\arguments{
\item{value}{a length 1 numeric vector}
}
\description{
An instance of an equal-to set
}
