% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result-status.R
\docType{data}
\name{MOI_FEASIBLE_POINT}
\alias{MOI_FEASIBLE_POINT}
\title{The feasible point result status code}
\format{
An object of class \code{MOI_feasible_point_result_status_code} of length 1.
}
\usage{
MOI_FEASIBLE_POINT
}
\description{
The feasible point result status code
}
\keyword{datasets}
