% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result-status.R
\docType{class}
\name{MOI_feasible_point_result_status_code-class}
\alias{MOI_feasible_point_result_status_code-class}
\title{The feasible point result status code class}
\description{
The feasible point result status code class
}
