% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_get,MOI_abstract_model,MOI_variable_primal_attribute,list-method}
\alias{moi_get,MOI_abstract_model,MOI_variable_primal_attribute,list-method}
\title{Get the primal value of a list of variables}
\usage{
\S4method{moi_get}{MOI_abstract_model,MOI_variable_primal_attribute,list}(model, type, index)
}
\arguments{
\item{model}{the model}

\item{type}{the variable primal attribute}

\item{index}{in this case a list of variables}
}
\description{
Get the primal value of a list of variables
}
