% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_get}
\alias{moi_get}
\title{Get something from a model}
\usage{
moi_get(model, type, index)
}
\arguments{
\item{model}{a MOI model}

\item{type}{the type to retrieve}

\item{index}{the index}
}
\description{
Get something from a model
}
