% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\docType{class}
\name{MOI_greater_than_set-class}
\alias{MOI_greater_than_set-class}
\title{An greater than or equal set}
\description{
An greater than or equal set
}
\section{Slots}{

\describe{
\item{\code{lower}}{A length-one numeric vector}
}}

