% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{moi_greater_than_set}
\alias{moi_greater_than_set}
\title{An instance of a greater than or equal set}
\usage{
moi_greater_than_set(lower)
}
\arguments{
\item{lower}{a length 1 numeric vector}
}
\description{
An instance of a greater than or equal set
}
