% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination-status.R
\docType{class}
\name{MOI_infeasible_termination_status_code-class}
\alias{MOI_infeasible_termination_status_code-class}
\title{Infeasible termination status code class}
\description{
Infeasible termination status code class
}
