% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\docType{data}
\name{moi_integer_set}
\alias{moi_integer_set}
\title{An instance of an integer set}
\format{
An object of class \code{MOI_integer_set} of length 1.
}
\usage{
moi_integer_set
}
\description{
An instance of an integer set
}
\keyword{datasets}
