% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\docType{class}
\name{MOI_interval_set-class}
\alias{MOI_interval_set-class}
\title{An interval set}
\description{
An interval set
}
\section{Slots}{

\describe{
\item{\code{lower}}{A length-one numeric vector}

\item{\code{upper}}{A length-one numeric vector}
}}

