% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\docType{class}
\name{MOI_less_than_set-class}
\alias{MOI_less_than_set-class}
\title{An less than or equal set}
\description{
An less than or equal set
}
\section{Slots}{

\describe{
\item{\code{upper}}{A length-one numeric vector}
}}

