% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-attributes.R
\docType{data}
\name{MOI_MIN_SENSE}
\alias{MOI_MIN_SENSE}
\title{Optimization min attribute}
\format{
An object of class \code{MOI_optimization_sense_min} of length 1.
}
\usage{
MOI_MIN_SENSE
}
\description{
Optimization min attribute
}
\keyword{datasets}
