% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-attributes.R
\docType{data}
\name{moi_number_of_constraints}
\alias{moi_number_of_constraints}
\title{Number of constraints attribute}
\format{
An object of class \code{MOI_number_of_constraints} of length 1.
}
\usage{
moi_number_of_constraints
}
\description{
Number of constraints attribute
}
\keyword{datasets}
