% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{class}
\name{MOI_objective_bound-class}
\alias{MOI_objective_bound-class}
\title{The objective bound attribute of an optimizer}
\description{
The objective bound attribute of an optimizer
}
