% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{data}
\name{moi_objective_bound}
\alias{moi_objective_bound}
\title{The objective bound attribute}
\format{
An object of class \code{MOI_objective_bound} of length 1.
}
\usage{
moi_objective_bound
}
\description{
The objective bound attribute
}
\keyword{datasets}
