% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{class}
\name{MOI_objective_value-class}
\alias{MOI_objective_value-class}
\title{The objective value attribute of an optimizer}
\description{
The objective value attribute of an optimizer
}
\section{Slots}{

\describe{
\item{\code{result_idx}}{the length 1 index}
}}

