% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\name{moi_objective_value}
\alias{moi_objective_value}
\title{The objective value attribute}
\usage{
moi_objective_value(result_idx = 1)
}
\arguments{
\item{result_idx}{a result index}
}
\description{
The objective value attribute
}
