% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-attributes.R
\docType{class}
\name{MOI_optimization_sense_feasibility-class}
\alias{MOI_optimization_sense_feasibility-class}
\title{Optimization feasibility sense attribute class}
\description{
Optimization feasibility sense attribute class
}
