% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination-status.R
\docType{class}
\name{MOI_optimize_not_called_termination_status_code-class}
\alias{MOI_optimize_not_called_termination_status_code-class}
\title{Optimize not called status code class}
\description{
Optimize not called status code class
}
