% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination-status.R
\docType{class}
\name{MOI_other_error_termination_status_code-class}
\alias{MOI_other_error_termination_status_code-class}
\title{Other error status code class}
\description{
Other error status code class
}
