% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\name{moi_primal_status}
\alias{moi_primal_status}
\title{Primate status attribute}
\usage{
moi_primal_status(result_idx = 1)
}
\arguments{
\item{result_idx}{a result index (scalar >= 1)}
}
\description{
Primate status attribute
}
