% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{data}
\name{moi_relative_gap}
\alias{moi_relative_gap}
\title{The relative GAP attribute}
\format{
An object of class \code{MOI_relative_gap} of length 1.
}
\usage{
moi_relative_gap
}
\description{
The relative GAP attribute
}
\keyword{datasets}
