% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{data}
\name{moi_result_count}
\alias{moi_result_count}
\title{The result count attribute}
\format{
An object of class \code{MOI_result_count} of length 1.
}
\usage{
moi_result_count
}
\description{
The result count attribute
}
\keyword{datasets}
