% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{MOI_scalar_affine_function-class}
\alias{MOI_scalar_affine_function-class}
\title{A scalar affine function}
\description{
A scalar affine function
}
\section{Slots}{

\describe{
\item{\code{terms}}{a list of terms (MOI_scalar_affine_term)}

\item{\code{constant}}{a length 1 numeric}
}}

