% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moi_scalar_affine_function}
\alias{moi_scalar_affine_function}
\title{A scalar affine function}
\usage{
moi_scalar_affine_function(terms, constant)
}
\arguments{
\item{terms}{a list of affine terms}

\item{constant}{a constant (scalar numeric)}
}
\description{
A scalar affine function
}
