% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{MOI_scalar_affine_term-class}
\alias{MOI_scalar_affine_term-class}
\title{A scalar affine time}
\description{
A scalar affine time
}
\section{Slots}{

\describe{
\item{\code{variable}}{a MOI Variable Index}

\item{\code{coefficient}}{a length 1 numeric}
}}

