% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moi_scalar_affine_term}
\alias{moi_scalar_affine_term}
\title{A scalar affine term}
\usage{
moi_scalar_affine_term(coefficient, variable)
}
\arguments{
\item{coefficient}{a scalar numeric coefficient}

\item{variable}{a MOI variable}
}
\description{
A scalar affine term
}
