% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{MOI_scalar_quadratic_function-class}
\alias{MOI_scalar_quadratic_function-class}
\title{A scalar quadratic function}
\description{
A scalar quadratic function
}
\section{Slots}{

\describe{
\item{\code{affine_terms}}{a list of affine terms (MOI_scalar_affine_term)}

\item{\code{quadratic_terms}}{a list of quadratic terms (MOI_scalar_quadratic_term)}

\item{\code{constant}}{a length 1 numeric}
}}

