% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moi_scalar_quadratic_function}
\alias{moi_scalar_quadratic_function}
\title{A scalar quadratic function}
\usage{
moi_scalar_quadratic_function(quadratic_terms, affine_terms, constant)
}
\arguments{
\item{quadratic_terms}{A list of quadratic terms}

\item{affine_terms}{A list of affine terms}

\item{constant}{a scalar numeric constant}
}
\description{
A scalar quadratic function
}
