% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{MOI_scalar_quadratic_term-class}
\alias{MOI_scalar_quadratic_term-class}
\title{A scalar quadratic term}
\description{
A scalar quadratic term
}
\section{Slots}{

\describe{
\item{\code{variable1}}{A MOI variable index}

\item{\code{variable2}}{A MOI variable index}

\item{\code{coefficient}}{a length 1 numeric}
}}

