% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moi_scalar_quadratic_term}
\alias{moi_scalar_quadratic_term}
\title{A scalar quadratic term}
\usage{
moi_scalar_quadratic_term(coefficient, variable1, variable2)
}
\arguments{
\item{coefficient}{a scalar numeric}

\item{variable1}{a variable}

\item{variable2}{another variable}
}
\description{
A scalar quadratic term
}
