% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_set}
\alias{moi_set}
\title{Modify a model}
\usage{
moi_set(model, type, index, value)
}
\arguments{
\item{model}{the model}

\item{type}{what you want to modify}

\item{index}{where you want to modify something}

\item{value}{the optional value}
}
\description{
Modify a model
}
