% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{MOI_single_variable-class}
\alias{MOI_single_variable-class}
\title{A single variable}
\description{
A single variable
}
\section{Slots}{

\describe{
\item{\code{variable}}{a MOI Variable Index}
}}

