% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moi_single_variable}
\alias{moi_single_variable}
\title{A single variable}
\usage{
moi_single_variable(variable_index)
}
\arguments{
\item{variable_index}{a scalar variable index}
}
\description{
A single variable
}
