% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination-status.R
\docType{class}
\name{MOI_success_termination_status_code-class}
\alias{MOI_success_termination_status_code-class}
\title{Success termination status code class}
\description{
Success termination status code class
}
