% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_supports,MOI_abstract_model,MOI_abstract_constraint_attribute-method}
\alias{moi_supports,MOI_abstract_model,MOI_abstract_constraint_attribute-method}
\title{Default implementation supporting constraint attributes}
\usage{
\S4method{moi_supports}{MOI_abstract_model,MOI_abstract_constraint_attribute}(model, type)
}
\arguments{
\item{model}{an abstract model}

\item{type}{the abstract constraint attribute
Should be overwritten.}
}
\description{
Default implementation supporting constraint attributes
}
