% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_supports,MOI_abstract_model,MOI_abstract_variable_attribute-method}
\alias{moi_supports,MOI_abstract_model,MOI_abstract_variable_attribute-method}
\title{Default implementation supporting variable attributes}
\usage{
\S4method{moi_supports}{MOI_abstract_model,MOI_abstract_variable_attribute}(model, type)
}
\arguments{
\item{model}{an abstract model}

\item{type}{the abstract variable attribute
Should be overwritten.}
}
\description{
Default implementation supporting variable attributes
}
