% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{moi_supports}
\alias{moi_supports}
\title{Check if a model supports something}
\usage{
moi_supports(model, type)
}
\arguments{
\item{model}{a MOI_model}

\item{type}{the type it supports}
}
\description{
Check if a model supports something
}
