% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{class}
\name{MOI_termination_message-class}
\alias{MOI_termination_message-class}
\title{The termination message attribute of an optimizer}
\description{
The termination message attribute of an optimizer
}
