% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{data}
\name{moi_termination_message}
\alias{moi_termination_message}
\title{Termination message attribute}
\format{
An object of class \code{MOI_termination_message} of length 1.
}
\usage{
moi_termination_message
}
\description{
Termination message attribute
}
\keyword{datasets}
