% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{data}
\name{moi_termination_solver_message}
\alias{moi_termination_solver_message}
\title{Termination solver message attribute}
\format{
An object of class \code{MOI_termination_solver_message_attribute} of length 1.
}
\usage{
moi_termination_solver_message
}
\description{
Termination solver message attribute
}
\keyword{datasets}
