% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{class}
\name{MOI_termination_solver_message_attribute-class}
\alias{MOI_termination_solver_message_attribute-class}
\title{The termination solver message attribute of an optimizer}
\description{
The termination solver message attribute of an optimizer
}
