% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{class}
\name{MOI_termination_status-class}
\alias{MOI_termination_status-class}
\title{The termination status attribute of an optimizer}
\description{
The termination status attribute of an optimizer
}
