% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-attributes.R
\docType{data}
\name{moi_termination_status}
\alias{moi_termination_status}
\title{Termination status attribute}
\format{
An object of class \code{MOI_termination_status} of length 1.
}
\usage{
moi_termination_status
}
\description{
Termination status attribute
}
\keyword{datasets}
