% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-attributes.R
\docType{data}
\name{moi_variable_dual}
\alias{moi_variable_dual}
\title{A variable dual attribute}
\format{
An object of class \code{MOI_variable_dual_attribute} of length 1.
}
\usage{
moi_variable_dual
}
\description{
A variable dual attribute
}
\keyword{datasets}
