% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indextypes.R
\docType{class}
\name{MOI_variable_index-class}
\alias{MOI_variable_index-class}
\title{A variable index}
\description{
A variable index
}
\section{Slots}{

\describe{
\item{\code{value}}{a length one integer like numeric}
}}

