% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-attributes.R
\docType{data}
\name{moi_variable_name}
\alias{moi_variable_name}
\title{A variable name attribute}
\format{
An object of class \code{MOI_variable_name_attribute} of length 1.
}
\usage{
moi_variable_name
}
\description{
A variable name attribute
}
\keyword{datasets}
