% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-attributes.R
\docType{data}
\name{moi_variable_primal}
\alias{moi_variable_primal}
\title{A variable primal attribute}
\format{
An object of class \code{MOI_variable_primal_attribute} of length 1.
}
\usage{
moi_variable_primal
}
\description{
A variable primal attribute
}
\keyword{datasets}
