% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-attributes.R
\docType{data}
\name{moi_variable_primal_start}
\alias{moi_variable_primal_start}
\title{A variable primal start attribute}
\format{
An object of class \code{MOI_variable_primal_start_attribute} of length 1.
}
\usage{
moi_variable_primal_start
}
\description{
A variable primal start attribute
}
\keyword{datasets}
