% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-attributes.R
\docType{class}
\name{MOI_variable_primal_start_attribute-class}
\alias{MOI_variable_primal_start_attribute-class}
\title{Variable primal start attribute}
\description{
Variable primal start attribute
}
