% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\docType{data}
\name{moi_zero_one_set}
\alias{moi_zero_one_set}
\title{An instance of a set of binaries}
\format{
An object of class \code{MOI_zero_one_set} of length 1.
}
\usage{
moi_zero_one_set
}
\description{
All elements must be integral and 0/1.
}
\keyword{datasets}
