% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result-status.R
\docType{data}
\name{NO_SOLUTION}
\alias{NO_SOLUTION}
\title{The no solution status code}
\format{
An object of class \code{MOI_no_solution_result_status_code} of length 1.
}
\usage{
NO_SOLUTION
}
\description{
The no solution status code
}
\keyword{datasets}
