% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-attributes.R
\docType{data}
\name{number_of_variables}
\alias{number_of_variables}
\title{Number of variables attribute}
\format{
An object of class \code{MOI_number_of_variables} of length 1.
}
\usage{
number_of_variables
}
\description{
Number of variables attribute
}
\keyword{datasets}
