% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-attributes.R
\docType{data}
\name{objective_function}
\alias{objective_function}
\title{Objective function attribute}
\format{
An object of class \code{MOI_objective_function} of length 1.
}
\usage{
objective_function
}
\description{
Objective function attribute
}
\keyword{datasets}
