% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-attributes.R
\docType{data}
\name{objective_sense}
\alias{objective_sense}
\title{Objective sense attribute}
\format{
An object of class \code{MOI_objective_sense} of length 1.
}
\usage{
objective_sense
}
\description{
Objective sense attribute
}
\keyword{datasets}
