% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination-status.R
\docType{data}
\name{MOI_OPTIMIZE_NOT_CALLED}
\alias{MOI_OPTIMIZE_NOT_CALLED}
\alias{MOI_OPTIMAL}
\alias{MOI_SUCCESS}
\alias{MOI_OTHER_ERROR}
\alias{MOI_INFEASIBLE}
\alias{MOI_INFEASIBLE_OR_UNBOUNDED}
\title{Status codes}
\format{
An object of class \code{MOI_optimize_not_called_termination_status_code} of length 1.

An object of class \code{MOI_optimal_termination_status_code} of length 1.

An object of class \code{MOI_success_termination_status_code} of length 1.

An object of class \code{MOI_other_error_termination_status_code} of length 1.

An object of class \code{MOI_infeasible_termination_status_code} of length 1.

An object of class \code{MOI_infeasible_or_unbounded_termination_status_code} of length 1.
}
\usage{
MOI_OPTIMIZE_NOT_CALLED

MOI_OPTIMAL

MOI_SUCCESS

MOI_OTHER_ERROR

MOI_INFEASIBLE

MOI_INFEASIBLE_OR_UNBOUNDED
}
\description{
Status codes
}
\keyword{datasets}
