% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination-status.R
\docType{data}
\name{OPTIMIZE_NOT_CALLED}
\alias{OPTIMIZE_NOT_CALLED}
\alias{OPTIMAL}
\alias{SUCCESS}
\alias{OTHER_ERROR}
\alias{INFEASIBLE}
\alias{INFEASIBLE_OR_UNBOUNDED}
\title{Status codes}
\format{
An object of class \code{MOI_optimize_not_called_termination_status_code} of length 1.

An object of class \code{MOI_optimal_termination_status_code} of length 1.

An object of class \code{MOI_success_termination_status_code} of length 1.

An object of class \code{MOI_other_error_termination_status_code} of length 1.

An object of class \code{MOI_infeasible_termination_status_code} of length 1.

An object of class \code{MOI_infeasible_or_unbounded_termination_status_code} of length 1.
}
\usage{
OPTIMIZE_NOT_CALLED

OPTIMAL

SUCCESS

OTHER_ERROR

INFEASIBLE

INFEASIBLE_OR_UNBOUNDED
}
\description{
Status codes
}
\keyword{datasets}
